+++ src/chains/ethereum/ethereum.config.ts
@@ -1,29 +1,31 @@
-type TokenListType = 'FILE' | 'URL';
+import { TokenListType } from '../../services/base';
 
-export interface Config {
-  chainId: number;
-  rpcUrl: string;
-  tokenListType: TokenListType;
-  tokenList: string;
+export namespace EthereumConfig {
+  export interface NetworkConfig {
+    chainId: number;
+    rpcUrl: string;
+    tokenListType: TokenListType;
+    tokenListSource: string;
+  }
+
+  export interface Config {
+    mainnet: NetworkConfig;
+    kovan: NetworkConfig;
+  }
+
+  export const config: Config = {
+    mainnet: {
+      chainId: 1,
+      rpcUrl: `https://mainnet.infura.io/v3/${process.env.INFURA_KEY}`,
+      tokenListType: 'URL',
+      tokenListSource:
+        'https://wispy-bird-88a7.uniswap.workers.dev/?url=http://tokens.1inch.eth.link',
+    },
+    kovan: {
+      chainId: 42,
+      rpcUrl: `https://kovan.infura.io/v3/${process.env.INFURA_KEY}`,
+      tokenListType: 'FILE',
+      tokenListSource: 'src/chains/ethereum/erc20_tokens_kovan.json',
+    },
+  };
 }
-
-export interface EthereumConfig {
-  mainnet: Config;
-  kovan: Config;
-}
-
-export const ethereumConfig: EthereumConfig = {
-  mainnet: {
-    chainId: 1,
-    rpcUrl: `https://mainnet.infura.io/v3/${process.env.INFURA_KEY}`,
-    tokenListType: 'URL',
-    tokenList:
-      'https://wispy-bird-88a7.uniswap.workers.dev/?url=http://tokens.1inch.eth.link',
-  },
-  kovan: {
-    chainId: 42,
-    rpcUrl: `https://kovan.infura.io/v3/${process.env.INFURA_KEY}`,
-    tokenListType: 'URL',
-    tokenList: '',
-  },
-};