+++ Changelog.md
@@ -4,6 +4,23 @@ All notable changes to gateway should be added to this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2021-07-27
+
+### Added
+
+- Tests for config-manager.ts.
+
+### Changed
+
+- Drop EthTransactionReceipt, use providers.TransactionReceipt for getTransactionReceipt.
+
+- Trade yaml for js-yaml because it has types.
+
+- Add a Config type to ConfigManager.
+
+- Ethereum extends EthereumBase and has custom behavior for the ethereum gas price.
+
+
 ## [0.1.1] - 2021-07-26
 
 ### Added